requirements.txt
# Dependências principais
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0

# Dependências de desenvolvimento (opcional)
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Dependências do RAG
chromadb>=0.4.20
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
pypdf>=4.0.0
